resolved complexity issues
